ft<>publish delete book message

+++ AdminService/app/services/book_service.py
@@ -44,6 +44,12 @@ class BookService:
             return result
     
     # DELETE A BOOK
-    async def delete_book(book_id):
-        await BookRepository.delete_book_by_id(book_id)
+    async def delete_book(book_id, producer:RabbitMQClient):
+        # Get book by id
+        book = await BookRepository.get_book_by_id(book_id)
+        print(book)
+        if book is not None:
+            message = json.dumps({"book_universal_id": book["universal_id"]}) 
+            await producer.publish("delete_book", message)
+            await BookRepository.delete_book_by_id(book_id)
         return
